Add adjustment to Recast Demo for TrinityCore MMAPs.
- enable DT_POLYREF64
- fix some Win x64 build and debug issues
- add TrinityCore types and defines
- allow *.mmtile files to be loaded and displayed through the "Load mmtile" button
- Sample_SoloMesh loads only the tile of the current *.obj
- Sample_TileMesh loads the tile of the current *.obj and its 8 neighbors if available
- speed up the mouse wheel zoom

+++ RecastDemo/Source/Filelist.cpp
@@ -65,7 +65,7 @@ void scanDirectory(const char* path, const char* ext, FileList& list)
 #ifdef WIN32
 	_finddata_t dir;
 	char pathWithExt[260];
-	long fh;
+    intptr_t fh;
 	strcpy(pathWithExt, path);
 	strcat(pathWithExt, "/*");
 	strcat(pathWithExt, ext);
